requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-xai>=0.1.0
streamlit>=1.28.0
tree-sitter>=0.20.0
tree-sitter-language-pack>=0.7.0
python-dotenv>=1.0.0